backend/requirements.txt
# FastAPI Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email
fastapi-mail==1.4.1
jinja2==3.1.3

# File Upload & Image Processing
cloudinary==1.39.0
python-magic==0.4.27
pillow==10.2.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Environment Variables
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

# Production Server
gunicorn==21.2.0
